Do not restart service at upgrade
Restarting the service can interrupt recordings or
current viewing.  An RFE has been submitted to provide
a way to determine if the record engine is currently
active, or to request it restart when it is next idle,
but until that is resolved do not automatically restart
the record engine at upgrade.  This will require the
user to manually restart the record engine when the
system is otherwise idle to gain changes/fixes.