eliminate use of mockchain for build of MythTV
The build script previously used mockchain, which
was itself problematic and required local fixes due
to open issues.  With recent mock versions, upstream
has closed the issues as they have determined that
the best way forward is to deprecate mockchain
itself and replaced it with the --chain option to
mock.  Unfortunately, --chain has its own set of
issues when utilizing a localrepo (the localrepo
repodata directory is recreated inside the chroot
as root, making the resulting temporary directory
protected against deletion at cleanup).  Instead
(at least for now) we will use mock itself by
performing some of the equivalent steps, at the
cost of not being able to exploit a root cache due
to some packages that are installed due to
dependency requirements that are pulled in being
marked as protected (there is fix for that) or not
being determined to be unneeded (so cleanup of the
root cache is incomplete even with when bypassing
the protected status), and just clean up the
root-cache as needed.