workaround bug/feature with mock --install for MythTV
For individuals that have not enabled the build
performance enhancements of using the tmpfs plugin
there appears to be a bug in mock where the
--cache-alternations option will not recreate the
root cache when using --install.  So we do the
--install twice, once for those without the tmpfs
plugin, and once for those with the tmpfs plugin
enabled.  I'll follow up with the mock project
with a smaller reproducer at some future point,
but this seems to work for now.