adjust -devel and -libs packages for MythTV
The -devel package should not install the
entire set of potential packages needed for
development (that is ansible's job), but
just the development libraries themselves
(typically the .so links, and the headers).
In practice, I suspect this would primarily
be used for development of plugins (which
has not actually happened in some time).