adjust flag processing for improved compatibility
Note that some of the flags are only passed along
with the upcoming MythTV 32 release (or the latest
development branch) due to FFmpeg flag processing
artifacts before the FFmpeg 4.3.1 merge

llvm 11 has not yet landed in Fedora 33, so the
test is anticipatory.

The change to use the llvm project lld linker when
using the clang compiler is necessary for a future
update.