Preliminary support for building CentOS 9 for MythTV
In addition to specifing the local ./build_cfg
for epel-9... as the --root for mock, one will
also need to also specify --without rpmfusion for
the build to be successful.  It is expected that
at least some of the workarounds can eventually
be removed when CentOS 9 is released and the
needed software is built.