revise build_cfgs for centos stream 8/9 for MythTV
With mock-core-configs 36.4 the layouts have changed
to reflect a new approach to the core and template
names, and the removal of the devel repo for
centos stream 8 by providing our own copr builds.
Note that while rawhide (i.e. to be F36) includes
mock-core-configs 36.4, this revision has (at this
time) not been pushed to the updates repo on
previous releases due to considerations about
centos stream and the development libraries and
alternative enterprise 8 distributions.  One can
get the latest mock-core-configs by:
  sudo dnf upgrade --enablerepo=updates-testing \
           --advisory=FEDORA-2021-a7d4aaa6fe