Remove unneeded perl binary compat require for MythTV
The rpm perl-generators have been updated to create
any necessary dependencies.  Note that pure-perl
libraries (such as provided in MythTV) do not need
to be tightly locked to the specific perl binary
version, just a dependency on perl-libs, which can
result in a reduction of the need to rebuild many
packages due to a perl version update.