EL7 is now EOL
Support for EL7 will be removed from the
various packages in this repository moving
forward to simplify the configurations
and build processes.

The mock configs for EL7 that many of
the various packages depended on for
building EL7 rpm have been removed
by the mock team.  The frozen EL7
packages have been moved to the vault.